ai_agents/agente_diversos/pyproject.toml
[project]
name = "agente-diversos"
version = "0.1.0"
description = ""
authors = [
    {name = "WandersonCorreia",email = "wandersoncorreia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "openai (>=1.109.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "agente_diversos", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
